Fix transpiling of func. definitions with parameters and one statement
According to the GLSL language specification, version 4.60, function
definitions are a function prototype followed by a compound statement,
which always includes braces. Previously we didn't include braces if
this compound statement only contained a single statement, which wasn't
correct.

Also, a space between a type and identifier is necessary in function
parameters.