Fix preprocessor directives not being put in separate lines sometimes
The GLSL specification, which refers to the C++ ISO/IEC 14882:1998
standard for preprocessor directive syntax, indicates that preprocessor
directives should belong in their own line.

Previously, if a preprocessor directive followed another external
declaration that doesn't have this requirement and didn't insert a
newline, like a variable declaration, the preprocessor directive would
be inserted in an existing line, which is incorrect.

Fix the problem by making the show_translation_unit function aware of
external declarations that require their own separate line, so it can
insert newlines as needed.